+++ __fixtures__/typescript/src/index.ts
@@ -1,3 +1,9 @@
-let thing: string = "something";
+import { SomeType } from "./another-thing";
+export * from "./one-more-thing";
+import * as path from "path";
+
+export { path };
+
+let thing: SomeType = "something";
 
 export default thing;