+++ packages/preconstruct/src/constants.js
@@ -1,5 +1,0 @@
-// @flow
-
-exports.EXTENSIONS = [".js", ".jsx", ".ts", ".tsx"];
-
-exports.PKG_JSON_CONFIG_FIELD = "preconstruct";