+++ packages/preconstruct/src/constants.ts
@@ -1,0 +1,3 @@
+export let EXTENSIONS = [".js", ".jsx", ".ts", ".tsx"] as const;
+
+export let PKG_JSON_CONFIG_FIELD = "preconstruct" as const;