+++ site/gatsby-config.js
@@ -1,7 +1,20 @@
 module.exports = {
-  __experimentalThemes: ["gatsby-theme-sidebar"],
   plugins: [
-    "gatsby-mdx",
+    "gatsby-theme-sidebar",
+    {
+      resolve: `gatsby-source-filesystem`,
+      options: {
+        name: `pages`,
+        path: `${__dirname}/src/pages/`
+      }
+    },
+    {
+      resolve: "gatsby-plugin-page-creator",
+      options: {
+        path: `${__dirname}/src/pages/`
+      }
+    },
+    "gatsby-plugin-mdx",
     {
       resolve: "gatsby-plugin-typography",
       options: {
@@ -13,6 +26,18 @@ module.exports = {
       resolve: "gatsby-plugin-google-analytics",
       options: {
         trackingId: "UA-140394521-1"
+      }
+    },
+    {
+      resolve: `gatsby-plugin-manifest`,
+      options: {
+        name: "Preconstruct",
+        short_name: "Preconstruct",
+        start_url: "/",
+        background_color: "#ea2929",
+        theme_color: "#ea2929",
+        display: "browser",
+        icon: "src/images/icon.png"
       }
     }
   ]